+++ R/globals.R
@@ -1,3 +1,5 @@
-utils::globalVariables(names = c(":=", ".N", ".SD", ".",
-                                 "get_SODB_dataset", "check_SODB_adata",
-                                  "list_SODB_datasets", "list_SODB_dataset_experiments"))
+utils::globalVariables(names = c(
+    ":=", ".N", ".SD", ".",
+    "get_SODB_dataset", "check_SODB_adata",
+    "list_SODB_datasets", "list_SODB_dataset_experiments")
+)